refactor: change error handler from use cases
