+++ src/common/errors/errors/already-exists-error.ts
@@ -1,7 +1,9 @@
+import { Result } from '@common/logic/result';
+
 import { UseCaseError } from '../use-case-error';
 
-export class AlreadyExistsError extends UseCaseError {
+export class AlreadyExistsError extends Result<UseCaseError> {
   constructor(value: string) {
-    super(`The ${value} already exists`);
+    super(false, { message: `The ${value} already exists` } as UseCaseError);
   }
 }
