+++ src/modules/project-management/application/use-cases/commands/comment-on-answer.ts
@@ -13,7 +13,7 @@ interface CommentOnAnswerUseCaseRequest {
 }
 
 type CommentOnAnswerUseCaseResponse = Either<
-  ResourceNotFoundError | Result<void>,
+  ResourceNotFoundError | Result<any>,
   Result<void>
 >;
