+++ src/modules/project-management/application/use-cases/commands/errors/manage-project-team-member-privilege-error.ts
